fix: rename type as types
