+++ packages/wallets/src/components/Base/WalletButton/WalletButton.tsx
@@ -1,6 +1,6 @@
 import React, { ButtonHTMLAttributes, ReactElement } from 'react';
 import classNames from 'classnames';
-import { TGenericSizes } from '../type';
+import { TGenericSizes } from '../types';
 import { WalletText } from '../WalletText';
 import styles from './WalletButton.module.css';
